Add Source property to Metar class and update database
Introduce a new Source property to the Metar class in the avwx_metar_service namespace. Update AppDbContextModelSnapshot.cs to include the new Source property in the Metar entity. Create a new migration file 20241205223419_addedSource.cs to handle the addition of the Source column to the Metars table in the database. Generate 20241205223419_addedSource.Designer.cs to reflect the changes in the database schema.